Remove unnecessary compatibility leftovers
